requirements.txt
asgiref==3.7.2
django==4.2.4
django-emoji==2.2.2
django-markdownx==4.0.2
markdown==3.4.4
Pillow==10.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
